Fix release and Wear CI gates
