+++ package.json
@@ -62,6 +62,7 @@
     "release:metadata": "node scripts/release-config.mjs metadata",
     "test:release": "node --test scripts/release-config.test.mjs scripts/dependency-advisory-exceptions.test.mjs",
     "test:native-release": "node --test scripts/native-release-build.test.mjs",
+    "test:mobile-build-tools": "node --test scripts/xcode-uuid-compatibility.test.mjs",
     "test:native:upgrade": "node scripts/native-upgrade-rehearsal.mjs",
     "test:native:upgrade:unit": "node --test scripts/native-upgrade-rehearsal.test.mjs",
     "test:build-budget": "node --test scripts/frontend-build-budget.test.mjs",
